+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Regenerated JSVM types to include methods from structs with single generic parameter.
 
+- Fixed `RateLimitRule.Audience` code comment ([#7098](https://github.com/pocketbase/pocketbase/pull/7098); thanks @iustin05).
+
+- Try to forward Apple OAuth2 POST redirect user's name so that it can be returned (and eventually assigned) with the success response of the all-in-one auth call ([#7090](https://github.com/pocketbase/pocketbase/issues/7090)).
+
 
 ## v0.29.2